latest version 01/06/21

+++ madkting/models/product_template.py
@@ -19,7 +19,7 @@ class ProductTemplate(models.Model):
     _inherit = 'product.template'
 
     @api.model
-    def mdk_create(self, product_data):
+    def mdk_create(self, product_data, id_shop=None):
         """
         TODO: Prices are defined by product not variant. This behavior may be emulated with discount rules in odoo
               at this moment there is only price definition by product
@@ -65,8 +65,11 @@ class ProductTemplate(models.Model):
         :return:
         :rtype: dict
         """
-
+        logger.info("### PRODUCT DATA ###")
+        logger.info(product_data)
         config = self.env['madkting.config'].get_config()
+        mapping = self.env['yuju.mapping.product']
+        products = self.env['product.product']
 
         if config and config.simple_description_enabled:
             product_name = product_data.get("name", "")
@@ -78,6 +81,12 @@ class ProductTemplate(models.Model):
                 "description_pickingin" : product_name
             })
 
+        is_multi_shop = False
+        if product_data.get('is_multi_shop'):
+            product_data.pop('is_multi_shop')
+            is_multi_shop = True
+
+        company_id = product_data.get('company_id', False)
         variation_attributes = product_data.pop('variation_attributes', None)
         variations = product_data.pop('variations', [])
         has_variations = True if variation_attributes else False
@@ -97,6 +106,12 @@ class ProductTemplate(models.Model):
 
         if product_data.get('cost'):
             product_data['standard_price'] = product_data.pop('cost', None)
+
+        if product_data.get('barcode'):
+            product_ids = self.env['product.product'].sudo().search([('barcode', '=', product_data.get('barcode', ''))], limit=1)
+            if product_ids.ids:
+                return results.error_result(code='duplicated_barcode',
+                                            description='El codigo de barras ya esta previamente registrado')
 
         if product_data.get('l10n_mx_edi_code_sat_id'):
             logger.debug('Product data contains l10n_mx_edi_code_sat_id')
@@ -118,19 +133,63 @@ class ProductTemplate(models.Model):
                 logger.debug('Remove l10n_mx_edi_code_sat_id from product data')
                 product_data.pop('l10n_mx_edi_code_sat_id')
 
+
         # create a product simple
         if not has_variations:
-            try:
-                new_product_simple = self.env['product.product'].create(product_data)
-            except Exception as ex:
-                logger.exception(ex)
-                return results.error_result(code='product_create_error',
-                                            description='Product couldn\'t be created because '
-                                                        'of the following exception: {}'.format(ex))
+            if id_shop:
+                id_product_madkting = product_data.get('id_product_madkting')
+                sku_product = product_data.get('default_code')
+                product_mapping_data = {     
+                    'product_id' : False, 
+                    'id_product_yuju' : id_product_madkting,
+                    'id_shop_yuju' : id_shop,
+                    'default_code' : sku_product,
+                    'state' : 'active'
+                }
+                product_ids = self.env['product.product'].search([('default_code', '=', sku_product)], limit=1)
+                if product_ids:
+                    product_mapping_data.update({'product_id' : product_ids.id})
+                    try:
+                        mapping.create_or_update_product_mapping(product_mapping_data)
+                    except Exception as ex:
+                        logger.exception(ex)
+                        return results.error_result(code='product_create_error',
+                                                    description='Product Mapping couldn\'t be created because '
+                                                                'of the following exception: {}'.format(ex))
+
+                else:
+                    try:
+                        new_product_simple = self.env['product.product'].create(product_data)
+                    except Exception as ex:
+                        logger.exception(ex)
+                        return results.error_result(code='product_create_error',
+                                                    description='Product couldn\'t be created because '
+                                                                'of the following exception: {}'.format(ex))
+                    else:
+                        try:
+                            product_mapping_data.update({'product_id' : new_product_simple.id})
+                            mapping.create_or_update_product_mapping(product_mapping_data)
+                        except Exception as ex:
+                            logger.exception(ex)
+                            """Si no se logra crear el mapeo, se elimina el producto que se acaba de crear"""
+                            new_product_simple.unlink()
+                            return results.error_result(code='product_create_error',
+                                                        description='Product mapping couldn\'t be created because '
+                                                                    'of the following exception: {}'.format(ex))
+                        else:                        
+                            # if stock:
+                            #    pass # TODO: implement initial stock functionality
+                            return results.success_result(data=new_product_simple.get_data_with_variations())         
             else:
-                # if stock:
-                #    pass # TODO: implement initial stock functionality
-                return results.success_result(data=new_product_simple.get_data_with_variations())
+                try:
+                    new_product_simple = self.env['product.product'].create(product_data)
+                except Exception as ex:
+                    logger.exception(ex)
+                    return results.error_result(code='product_create_error',
+                                                description='Product couldn\'t be created because '
+                                                            'of the following exception: {}'.format(ex))
+                else:
+                    return results.success_result(data=new_product_simple.get_data_with_variations())
 
         # create product with variations
         # validate variations
@@ -184,13 +243,37 @@ class ProductTemplate(models.Model):
 
         product_data['attribute_line_ids'] = product_template_attribute_lines
         product_data.pop('id_product_madkting', None)
-        try:
-            new_template = self.create(product_data)
-        except Exception as ex:
-            logger.exception(ex)
-            return results.error_result(code='product_template_create_error',
-                                        description='Product couldn\'t be created because '
-                                                    'of the following exception: {}'.format(ex))
+        
+        if id_shop:
+            new_template = None            
+            sku_product = product_data.get('default_code')
+            product_ids = products.search([('default_code', '=', sku_product)], limit=1)
+            if product_ids:
+                new_template = product_ids.product_templ_id
+            else:
+                for var in variations:
+                    sku_product = var.get('default_code')
+                    product_ids = products.search([('default_code', '=', sku_product)], limit=1)
+                    if product_ids:
+                        new_template = product_ids.product_tmpl_id
+                        break
+
+            if not new_template:
+                try:
+                    new_template = self.create(product_data)
+                except Exception as ex:
+                    logger.exception(ex)
+                    return results.error_result(code='product_template_create_error',
+                                                description='Product couldn\'t be created because '
+                                                            'of the following exception: {}'.format(ex))
+        else:
+            try:
+                new_template = self.create(product_data)
+            except Exception as ex:
+                logger.exception(ex)
+                return results.error_result(code='product_template_create_error',
+                                            description='Product couldn\'t be created because '
+                                                        'of the following exception: {}'.format(ex))
         
         for product_variant in new_template.product_variant_ids:
             data = product_variant.get_data()
@@ -216,11 +299,34 @@ class ProductTemplate(models.Model):
                 variation_data.pop('product_id', None)
                 variation_data.pop('id', None)
 
+                if id_shop:
+                    is_multi_shop = False
+                    if variation_data.get('is_multi_shop'):
+                        variation_data.pop('is_multi_shop')
+                        is_multi_shop = True
+                        
+                    id_product_madkting = variation_data.get('id_product_madkting')
+                    sku_variant = variation_data.get('default_code')
+                    variant_mapping_data = {     
+                        'product_id' : product_variant.id,
+                        'id_product_yuju' : id_product_madkting,
+                        'id_shop_yuju' : id_shop,
+                        'default_code' : sku_variant,
+                        'state' : 'active'
+                    }                                               
+                    try:
+                        mapping.create_or_update_product_mapping(variant_mapping_data)
+                    except Exception as ex:
+                        logger.exception(ex)
+                        return results.error_result(code='product_create_error',
+                                                description='Product mapping couldn\'t be created because '
+                                                            'of the following exception: {}'.format(ex))
+                
                 product_variant.write(variation_data)
-        
+
         return results.success_result(new_template.product_variant_id.get_data_with_variations())
 
-    def change_product_status(self, template_id, active):
+    def change_product_status(self, template_id, active, id_shop=None):
         """
         :param template_id:
         :type template_id: int
@@ -237,7 +343,37 @@ class ProductTemplate(models.Model):
                 'The product that you are trying to change doesn\'t exists or has been deleted'
             )
         try:
-            product.active = active
+            if id_shop:
+                yuju_mapping = self.env['yuju.mapping'].sudo()
+                product_mapping = self.env['yuju.mapping.product'].sudo()
+                mapping = yuju_mapping.search([('id_shop_yuju', '=', id_shop)])
+                if not mapping.ids:
+                    return results.error_result(
+                        'product_not_found',
+                        'Mapping record not found for id_shop {} in activate/deactivate'.format(id_shop)
+                    )
+                mapping_state = 'active' if active else 'disabled'
+                update_template = True
+                for variation in product.product_variant_ids:
+                    if update_template:
+                        mapping_product = product_mapping.get_product_mapping_by_product(variation.id, only_active=True)
+                        if mapping_product and len(mapping_product.ids) > 1:
+                            update_template = False
+
+                if update_template:
+                    for variation in product.product_variant_ids:
+                        mapping_product = self.env['yuju.mapping.product'].sudo().search([('product_id', '=', variation.id), ('id_shop_yuju', '=', id_shop)], limit=1)
+                        if mapping_product.ids:
+                            mapping_product.write({'state' : mapping_state})                            
+                    product.active = active                
+                else:
+                    for variation in product.product_variant_ids:
+                        mapping_product = self.env['yuju.mapping.product'].sudo().search([('product_id', '=', variation.id), ('id_shop_yuju', '=', id_shop)], limit=1)
+                        if mapping_product.ids:
+                            mapping_product.write({'state' : mapping_state})
+
+            else:
+                product.active = active
         except Exception as ex:
             logger.exception(ex)
             return results.error_result('activate_product_error', str(ex))
@@ -247,7 +383,7 @@ class ProductTemplate(models.Model):
             )
 
     @api.model
-    def deactivate_product(self, template_id):
+    def deactivate_product(self, template_id, id_shop=None):
         """
         :param template_id:
         :type template_id: int
@@ -257,19 +393,19 @@ class ProductTemplate(models.Model):
         variant deactivation may bring adverse results,
         you should validate the default odoo behavior before allow this functionality
         """
-        return self.change_product_status(template_id, active=False)
+        return self.change_product_status(template_id, active=False, id_shop=id_shop)
 
     @api.model
-    def activate_product(self, template_id):
+    def activate_product(self, template_id, id_shop=None):
         """
         :param template_id:
         :type template_id: int
         :return:
         """
-        return self.change_product_status(template_id, active=True)
+        return self.change_product_status(template_id, active=True, id_shop=id_shop)
 
     @api.model
-    def delete_product(self, template_id):
+    def delete_product(self, template_id, id_shop):
         """
         :param template_id:
         :type template_id: int
@@ -284,7 +420,31 @@ class ProductTemplate(models.Model):
                 'The product that you are trying to delete doesn\'t exists or is deleted already'
             )
         try:
-            product.unlink()
+            if id_shop:
+                yuju_mapping = self.env['yuju.mapping'].sudo()
+                product_mapping = self.env['yuju.mapping.product'].sudo()
+                mapping = yuju_mapping.search([('id_shop_yuju', '=', id_shop)])
+                if not mapping.ids:
+                    return results.error_result(
+                        'product_not_found',
+                        'Mapping record not found for id_shop {}'.format(id_shop)
+                    )
+                delete_template = True
+                for variation in product.product_variant_ids:
+                    if delete_template:
+                        mapping_product = product_mapping.get_product_mapping_by_product(variation.id)
+                        if mapping_product and len(mapping_product.ids) > 1:
+                            delete_template = False
+                if delete_template:
+                    product.unlink() 
+                else:
+                    for variation in product.product_variant_ids:         
+                        mapping_product = product_mapping.get_product_mapping(variation.id, id_shop)
+                        if mapping_product:
+                            mapping_product.unlink()
+                               
+            else:
+                product.unlink()
         except (exceptions.ValidationError, psycopg2.IntegrityError) as ve:
             logger.exception(ve)
             return results.error_result(
